requirements.txt
kafka-python
redis

